src/requirements/dev.txt
-r base.txt

django-debug-toolbar==1.3.2